django-ussd/chains/requirements.txt
amqp==2.2.1
billiard==3.5.0.3
celery==4.1.0
certifi==2017.7.27.1
chardet==3.0.4
Django==1.11.4
django-annoying==0.10.3
djangorestframework==3.6.3
idna==2.5
Jinja2==2.9.6
kombu==4.1.0
MarkupSafe==1.0
PyConfigure==0.5.9
PyStaticConfiguration==0.10.3
pytz==2017.2
PyYAML==3.12
requests==2.18.3
six==1.10.0
structlog==17.2.0
urllib3==1.22
ussd-airflow==0.0.4.7
vine==1.1.4
